\name{nameswomengal16}
\alias{nameswomengal16}
\docType{data}
\title{ nameswomengal16 data
}
\description{
This dataset corresponds to 25 most frequent women's names by municipality in Galicia in 2016.
}
\usage{
data(nameswomengal16)
}
\format{
  \code{nameswomengal16} is a data frame with women's names from Galicia in 2016.
}

\source{
The data corresponds to 25 most frequent women's names by municipality in Galicia in 2016. 
The dataset contains 6 columns, \code{prov}: the province, \code{muni}: the municipality, \code{namuni}: the name of the municipality, \code{name}: the name, \code{number}: the number of people with that name and \code{population}: the total population considered by municipality.

These data have been extracted from the website of the Galician Institute of Statistics (IGE). The IGE offers information on the surnames and names of the population whose residence is in the Autonomous Community of Galicia. 
The base information for the elaboration data is the file of the Municipal Register of inhabitants of 2014 that the National Institute of Statistics (INE) provides to the IGE.
}
\references{
Galician Institute of Statistics (IGE), \href{https://www.ige.gal/igebdt/esq.jsp?idioma=gl&ruta=onomast/notamet.jsp}{https://www.ige.eu/}
}
\examples{
data(nameswomengal16)
}
\keyword{datasets}